feat(security): 实现基于 Redis 的 Token 管理机制- 新增 RedisTokenManager 类，用于生成、解析、校验、刷新 JWT Token - 添加 OnlineLoginUser 和 AuthenticationToken 模型类 - 修改 SecurityConfig，集成 TokenAuthenticationFilter - 更新 SysAuthController 和 SysAuthService 接口，支持刷新 Token 功能 - 重构 LoginController，改名为 SysAuthController - 调整 SysLoginService，改名为 SysAuthService - 新增 SysAuthServiceImpl 实现类 - 更新 SecurityUtils，添加获取当前请求 Token 的方法

+++ chuang-framework/src/main/java/cn/zhangchuangla/infrastructure/web/service/impl/TokenServiceImpl.java
@@ -102,7 +102,7 @@ public class TokenServiceImpl implements TokenService {
      * @param sysUserDetails 登录信息
      */
     public void refreshToken(SysUserDetails sysUserDetails) {
-        Long expire = securityProperties.getExpire();
+        Integer expire = securityProperties.getExpire();
         sysUserDetails.setLoginTime(System.currentTimeMillis());
         //fixme 待优化，过期时间优化
         sysUserDetails.setExpireTime(sysUserDetails.getLoginTime() + expire * MILLIS_MINUTE);
